figure-config/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
duckdb==1.1.2
sqlalchemy==2.0.36
requests==2.32.3
pyyaml==6.0.2
click==8.1.7
pydantic==2.9.2
python-multipart==0.0.12
watchdog==5.0.3
pytest==8.3.3